requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.3.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap5==0.7
dj-database-url
Django==4.2.7
django-crispy-forms==2.1
django-environ==0.11.2
django-extensions==3.2.3
django-redis==6.0.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
fpdf2
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
mercadopago==2.2.1
openpyxl==3.1.2
packaging==25.0
Pillow==10.1.0
prompt_toolkit==3.0.52
psycopg2-binary
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.1
reportlab==4.0.7
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint